Unzip DONE & Full dupes control

+++ server/routes/file.route.js
@@ -4,7 +4,7 @@ const passportjwt = require('passport-jwt');
 const asyncHandler = require('express-async-handler');
 const fileCtrl = require('../controllers/file.controller');
 const persmissionCtrl = require('../controllers/filePermissions.controller');
-var formidable = require('formidable');
+const formidable = require('formidable');
 const jwtDecode = require("jwt-decode");
 const router = express.Router();
 const User = require("../models/user.model");
@@ -109,7 +109,8 @@ async function insert(req, res) {
         let workDone;
 
         //File can be written here
-        baseDir = __dirname+"/../userDirectory/"+filePath+(unzipFiles?"/"+incomingFiles[0].name.split('.').splice(0,incomingFiles[0].name.split('.').length-1).join('.'):"")
+        let uploadedDir = unzipFiles?incomingFiles[0].name.split('.').splice(0,incomingFiles[0].name.split('.').length-1).join('.'):"";
+        baseDir = __dirname+"/../userDirectory/"+filePath+(unzipFiles?"/"+uploadedDir:"")
 
         // Subfolder based on zip file name
         if(unzipFiles) {
@@ -139,7 +140,7 @@ async function insert(req, res) {
 
                         //Fixing filename for creating dirs
                         zippedFileName = zippedFileName.substr(0,zippedFileName.length-1);
-                        var destPath = filePath+"/"+zippedFileName.split('/').splice(0,zippedFileName.split('/').length-1).join('/');
+                        var destPath = filePath+"/"+uploadedDir+"/"+zippedFileName.split('/').splice(0,zippedFileName.split('/').length-1).join('/');
                         if(destPath.endsWith('/')) { destPath = destPath.substr(0, destPath.length-1); }
                         zippedFileName = zippedFileName.split('/')[zippedFileName.split('/').length-1];
 
@@ -152,7 +153,8 @@ async function insert(req, res) {
                       else {
                         zip.file(zippedFileName).async('nodebuffer').then( // Is a file
                           (unzippedFileContents) => {
-                            promises.push(new Promise(function(resolve, reject) { createFileFromZip(unzippedFileContents, zippedFileName, baseDir, subDirPath, subDirName, userId, resolve, reject);}));
+
+                            promises.push(new Promise(function(resolve, reject) { createFileFromZip(unzippedFileContents, zippedFileName, baseDir, subDirPath, subDirName+"/"+uploadedDir, userId, resolve, reject);}));
                             if(i == Object.keys(contents.files).length-1) {
                               resolveParent(0);
                             }
@@ -233,7 +235,8 @@ async function getFileListByUserId(req, res) {
   });
 }
 
-async function createDirectory(dirPath, filePathDepth,dirName, baseDir, userId, unzipFiles, resolve, reject) {
+async function createDirectory(dirPath, filePathDepth, dirName, baseDir, userId, unzipFiles, resolve, reject) {
+  
   let folder = {
     'name': dirName,
     'path': "/"+dirPath,
@@ -242,14 +245,17 @@ async function createDirectory(dirPath, filePathDepth,dirName, baseDir, userId, 
 
   // Creating dir on file system
   if(unzipFiles) {
-    dirPath = dirPath.split('/').splice(filePathDepth, dirPath.split('/').length-filePathDepth).join('/');
+    dirPath = dirPath.split('/').splice(filePathDepth+1, dirPath.split('/').length-2).join('/');
   }
 
+    
   try {
+    if(fs.existsSync(baseDir+(unzipFiles?(dirPath.length>0?"/"+dirPath:"")+"/"+dirName:""))) { reject("Directory already exists"); return;}
     fs.ensureDirSync(baseDir+(unzipFiles?(dirPath.length>0?"/"+dirPath:"")+"/"+dirName:""));
   }
   catch (err) {
     reject(err);
+    return;
   }
 
   var folderId;
@@ -278,7 +284,6 @@ async function createDirectory(dirPath, filePathDepth,dirName, baseDir, userId, 
                   'file': insertedFolder,
                   'perm': permissionToCreate
                 }
-                console.log("resolved");
                 resolve(craftedResponse);
             }
           );
@@ -298,6 +303,7 @@ async function createFile(file, filePath, subdirPath, subdirName, userId, resolv
   }
   catch(err) {
     reject(err);
+    return;
   }
 
   //Adding perm to DB
@@ -348,10 +354,12 @@ async function createFile(file, filePath, subdirPath, subdirName, userId, resolv
 async function createFileFromZip(file, fileName, filePath, subdirPath, subdirName, userId, resolve, reject) {
 
   try {
+    if(fs.existsSync(filePath+"/"+fileName)) { reject("File already exists"); return; }
     fs.writeFileSync(filePath+"/"+fileName, file);
   }
   catch(err) {
     reject(err);
+    return;
   }
 
   var subdirFileName = fileName.split('/').splice(fileName.split('/').length-1,1).join('/');
@@ -360,7 +368,7 @@ async function createFileFromZip(file, fileName, filePath, subdirPath, subdirNam
   //Adding perm to DB
   let fileToUpload = {
     'name': subdirFileName,
-    'path': (subdirPath.length>0?"/":"")+subdirPath+"/"+subdirName+(subdirFilePath.length>0?"/"+subdirFilePath:""),
+    'path': ((subdirPath.length>0?"/":"")+subdirPath+"/"+subdirName+(subdirFilePath.length>0?"/"+subdirFilePath:"")),
     'type': 'f'
   };
